+++ crates/breez-sdk/core/src/persist/mysql/mod.rs
@@ -11,10 +11,8 @@ mod base;
 mod storage;
 
 // Re-export public configuration types and functions (with UniFFI annotations).
-// `PoolQueueMode` is also re-exported so consumers can construct `MysqlStorageConfig`
-// without depending on `spark_mysql` directly.
 #[allow(unused_imports)]
-pub use base::{MysqlStorageConfig, PoolQueueMode, default_mysql_storage_config};
+pub use base::{MysqlStorageConfig, default_mysql_storage_config};
 
 // Re-export pool factory and store factories
 pub(crate) use base::{create_mysql_token_store, create_mysql_tree_store, create_pool};
