+++ crates/spark-mysql/src/lib.rs
@@ -18,14 +18,14 @@ pub mod pool;
 mod token_store;
 mod tree_store;
 
-pub use config::{MysqlStorageConfig, PoolQueueMode, default_mysql_storage_config};
+pub use config::{MysqlStorageConfig, default_mysql_storage_config};
 pub use error::MysqlError;
 pub use token_store::{
     MysqlTokenStore, create_mysql_token_store, create_mysql_token_store_from_pool,
 };
 pub use tree_store::{MysqlTreeStore, create_mysql_tree_store, create_mysql_tree_store_from_pool};
 
-pub use migrations::run_migrations;
+pub use migrations::{Migration, run_migrations};
 pub use pool::{create_pool, map_db_error};
 
 pub use mysql_async;
